requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1 